+++ lib/src/barback/asset_environment.dart
@@ -10,7 +10,9 @@ import 'package:path/path.dart' as path;
 import 'package:watcher/watcher.dart';
 
 import '../cached_package.dart';
+import '../compiler.dart';
 import '../dart.dart' as dart;
+import '../dartdevc/dartdevc_environment.dart';
 import '../entrypoint.dart';
 import '../io.dart';
 import '../log.dart' as log;
@@ -18,10 +20,8 @@ import '../package.dart';
 import '../package_graph.dart';
 import '../source/cached.dart';
 import '../utils.dart';
-import 'dartdevc/dartdevc_environment.dart';
 import 'admin_server.dart';
 import 'barback_server.dart';
-import 'compiler.dart';
 import 'dart_forwarding_transformer.dart';
 import 'dart2js_transformer.dart';
 import 'load_all_transformers.dart';
