+++ lib/src/compiler.dart
