+++ lib/src/dartdevc/module_computer.dart
@@ -11,9 +11,9 @@ import 'package:barback/barback.dart';
 import 'package:path/path.dart' as p;
 
 import 'module.dart';
-import '../../barback.dart';
-import '../../io.dart';
-import '../../dart.dart' show isEntrypoint, isPart;
+import '../barback.dart';
+import '../io.dart';
+import '../dart.dart' show isEntrypoint, isPart;
 
 /// There are two "types" of modules, `public` and `private`.
 ///
