+++ lib/src/dartdevc/module_reader.dart
