+++ test/dartdevc/build_test.dart
@@ -2,11 +2,11 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:pub/src/barback/compiler.dart';
+import 'package:pub/src/compiler.dart';
 import 'package:scheduled_test/scheduled_test.dart';
 
-import '../../descriptor.dart' as d;
-import '../../test_pub.dart';
+import '../descriptor.dart' as d;
+import '../test_pub.dart';
 
 main() {
   integration("pub build --compiler=dartdevc creates all required sources", () {
