+++ test/dartdevc/summaries_test.dart
@@ -5,11 +5,11 @@
 import 'package:analyzer/src/summary/idl.dart';
 import 'package:scheduled_test/scheduled_test.dart';
 
-import 'package:pub/src/barback/dartdevc/summaries.dart';
+import 'package:pub/src/dartdevc/summaries.dart';
 
-import '../../descriptor.dart' as d;
-import '../../test_pub.dart';
-import '../../serve/utils.dart';
+import '../descriptor.dart' as d;
+import '../test_pub.dart';
+import '../serve/utils.dart';
 
 main() {
   integration(
@@ -79,6 +79,60 @@ void main() {}
     requestShould404('packages/foo/invalid$linkedSummaryExtension');
     endPubServe();
   });
+
+  integration(
+      "can output unlinked analyzer summaries for modules under lib and web",
+      () {
+    d.dir("foo", [
+      d.libPubspec("foo", "1.0.0"),
+      d.dir("lib", [
+        d.file(
+            "foo.dart",
+            """
+    void foo() {}
+    """)
+      ]),
+    ]).create();
+
+    d.dir(appPath, [
+      d.appPubspec({
+        "foo": {"path": "../foo"}
+      }),
+      d.dir("lib", [
+        d.file(
+            "hello.dart",
+            """
+  import 'package:foo/foo.dart';
+
+  hello() => 'hello';
+  """)
+      ]),
+      d.dir("web", [
+        d.file(
+            "main.dart",
+            """
+  import 'package:myapp/hello.dart';
+
+  void main() {}
+  """)
+      ])
+    ]).create();
+
+    pubGet();
+    pubServe(args: ['--compiler', 'dartdevc']);
+
+    unlinkedSummaryRequestShouldSucceed(
+        'web__main$unlinkedSummaryExtension', [endsWith('web/main.dart')]);
+    unlinkedSummaryRequestShouldSucceed(
+        'packages/myapp/lib__hello$unlinkedSummaryExtension',
+        [equals('package:myapp/hello.dart')]);
+    unlinkedSummaryRequestShouldSucceed(
+        'packages/foo/lib__foo$unlinkedSummaryExtension',
+        [equals('package:foo/foo.dart')]);
+    requestShould404('invalid$unlinkedSummaryExtension');
+    requestShould404('packages/foo/invalid$unlinkedSummaryExtension');
+    endPubServe();
+  });
 }
 
 void linkedSummaryRequestShouldSucceed(String uri,
@@ -93,5 +147,15 @@ void linkedSummaryRequestShouldSucceed(String uri,
         .map((info) => info.summaryPath)
         .where((path) => path.isNotEmpty);
     expect(summaryDepPaths, unorderedMatches(expectedSummaryDeps));
+  });
+}
+
+void unlinkedSummaryRequestShouldSucceed(
+    String uri, List<Matcher> expectedUnlinkedUris) {
+  var expected = unorderedMatches(expectedUnlinkedUris);
+  scheduleRequest(uri).then((response) {
+    expect(response.statusCode, 200);
+    var bundle = new PackageBundle.fromBuffer(response.bodyBytes);
+    expect(bundle.unlinkedUnitUris, expected);
   });
 }
