+++ test/dartdevc/util.dart
@@ -7,8 +7,8 @@ import 'dart:convert';
 import 'package:barback/barback.dart';
 import 'package:test/test.dart';
 
-import 'package:pub/src/barback/dartdevc/module.dart';
-import 'package:pub/src/barback/dartdevc/module_reader.dart';
+import 'package:pub/src/dartdevc/module.dart';
+import 'package:pub/src/dartdevc/module_reader.dart';
 
 import 'package:pub/src/io.dart' as io_helpers;
